improved error report in IupLua console
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@3253 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/srclua5/iuplua_getcolor.c
@@ -17,8 +17,8 @@
  
 static int GetColor(lua_State *L)
 {
-  int x = luaL_checkinteger(L,1);
-  int y = luaL_checkinteger(L,2);
+  int x = (int)luaL_checkinteger(L,1);
+  int y = (int)luaL_checkinteger(L, 2);
   unsigned char r = (unsigned char) luaL_optnumber(L,3,0);
   unsigned char g = (unsigned char) luaL_optnumber(L,4,0);
   unsigned char b = (unsigned char) luaL_optnumber(L,5,0);
